+++ Dangerfile.swift
@@ -8,23 +8,35 @@ let danger = Danger()
 // fileImport: DangerfileExtensions/ChangelogCheck.swift
 checkChangelog()
 
-print("Hi from the runtime")
-
-if (danger.git.createdFiles.count + danger.git.modifiedFiles.count - danger.git.deletedFiles.count > 10) {
+if (danger.git.createdFiles.count + danger.git.modifiedFiles.count - danger.git.deletedFiles.count > 300) {
     warn("Big PR, try to keep changes smaller if you can")
 }
 
-if danger.github.pullRequest.title.contains("WIP") {
-    warn("PR is classed as Work in Progress")
+let swiftFilesWithCopyright = danger.git.createdFiles.filter {
+    $0.fileType == .swift
+    && danger.utils.readFile($0).contains("//  Created by")
 }
 
-print("HI")
-_ = danger.github.api.me { response in
-    print("OK")
-    switch response {
-    case .success(let user):
-        message(user.name ?? "")
-    case .failure:
-        break
+if swiftFilesWithCopyright.count > 0 {
+    let files = swiftFilesWithCopyright.joined(separator: ", ")
+    warn("In Danger JS we don't include copyright headers, found them in: \(files)")
+}
+
+// Support running via `danger local`
+if (danger.github != nil) {
+    // These checks only happen on a PR
+    if danger.github.pullRequest.title.contains("WIP") {
+        warn("PR is classed as Work in Progress")
+    }
+
+    // TODO: We're still figuring this out
+    _ = danger.github.api.me { response in
+        print("OK")
+        switch response {
+        case .success(let user):
+            message(user.name ?? "")
+        case .failure:
+            break
+        }
     }
 }