+++ Sources/Danger/Settings.swift
@@ -1,10 +1,3 @@
-//
-//  Settings.swift
-//  Danger
-//
-//  Created by Franco Meloni on 06/11/2018.
-//
-
 struct Settings: Decodable {
     let github: GitHubSettings
 }