+++ Tests/DangerTests/DangerDSLResources/DangerDSLGitHub.swift
@@ -1,10 +1,3 @@
-//
-//  DangerDSLGitHub.swift
-//  Danger
-//
-//  Created by Franco Meloni on 06/11/2018.
-//
-
 public let DSLGitHubJSON = """
 
 {